+++ src/Vishva.ts
@@ -116,7 +116,7 @@ import { RuntimeRangeSharingEntry, restoreSharedSkeletonAnimations, deduplicateR
  */
 export class Vishva {
 
-    static version: string = "0.4.0-alpha.49";
+    static version: string = "0.4.0-alpha.50";
 
     public static worldName: string;